# docs/api/swagger.yaml
basePath: /
definitions:
...
  Failure-string:
    properties:
      code:
        type: integer
      error:
        type: string
    type: object
...
  Success-dto_JoinSideResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/dto.JoinSideResponse'
    type: object
...
  dto.JoinSideResponse:
    properties:
      membershipID:
        type: string
      sideID:
        type: string
    type: object
...
info:
  contact: {}
  description: This is api for upside application
  title: Upside API
  version: "1.0"
paths:
...
  /sides/{sideID}/_join:
    post:
      parameters:
      - description: side id
        in: path
        name: sideID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Success-dto_JoinSideResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Failure-string'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Failure-string'
      security:
      - BearerAuth: []
      summary: Join Side
      tags:
      - Sides
...
securityDefinitions:
  BearerAuth:
    description: 'Bearer token authentication schema, format: Bearer <your-credential-token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"